Initial setup scripts for a sqlite3 database.
create_tables is responsible for defining our tables (big shocker)
test_data populates those tables for the purposes of testing.

At some point it would be interesting to write a program that can create
the desired amount of fake data. This might exist, but it would be fun
to make.